Begin movies portion.
Setup routing.
Formatting.
Add basic routes to check working.
