+++ web/README.md
@@ -1,1 +1,6 @@
-# movies-ngrx
+# NgRx Movies Directory
+
+- Introduction to NgRx.
+- Thanks to [TMDB](https://www.themoviedb.org) for providing the API.
+- Following NgRx best practises as explained [here](https://indepth.dev/posts/1451/ngrx-best-practices-new).
+- Following Angular Style Guide as explained [here](https://angular.io/guide/styleguide#angular-ngmodule-names).
