+++ web/package.json
@@ -11,15 +11,16 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^12.2.11",
-        "@angular/common": "^12.2.11",
-        "@angular/compiler": "^12.2.11",
-        "@angular/core": "^12.2.11",
-        "@angular/forms": "^12.2.11",
-        "@angular/platform-browser": "^12.2.11",
-        "@angular/platform-browser-dynamic": "^12.2.11",
-        "@angular/router": "^12.2.11",
+        "@angular/animations": "^12.2.12",
+        "@angular/common": "^12.2.12",
+        "@angular/compiler": "^12.2.12",
+        "@angular/core": "^12.2.12",
+        "@angular/forms": "^12.2.12",
+        "@angular/platform-browser": "^12.2.12",
+        "@angular/platform-browser-dynamic": "^12.2.12",
+        "@angular/router": "^12.2.12",
         "@ngrx/store": "^12.5.1",
+        "@ngrx/store-devtools": "^12.5.1",
         "rxjs": "^7.4.0",
         "tslib": "^2.3.1",
         "zone.js": "^0.11.4"
@@ -31,12 +32,12 @@
         "@angular-eslint/eslint-plugin-template": "12.3.1",
         "@angular-eslint/schematics": "12.3.1",
         "@angular-eslint/template-parser": "12.3.1",
-        "@typescript-eslint/eslint-plugin": "4.28.5",
-        "@typescript-eslint/parser": "4.28.5",
         "@angular/cli": "^12.2.11",
-        "@angular/compiler-cli": "^12.2.11",
+        "@angular/compiler-cli": "^12.2.12",
         "@types/jasmine": "^3.10.1",
         "@types/node": "^16.11.6",
+        "@typescript-eslint/eslint-plugin": "4.28.5",
+        "@typescript-eslint/parser": "4.28.5",
         "codelyzer": "^6.0.2",
         "eslint": "^7.32.0",
         "eslint-plugin-import": "^2.25.2",
