--- web/src/app/app.component.html
@@ -1,1 +1,0 @@
-<movies-my-counter></movies-my-counter>
