--- web/src/app/app.module.ts
@@ -1,24 +1,0 @@
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-
-import {AppComponent} from './app.component';
-import {StoreModule} from "@ngrx/store";
-import {counterReducer} from "./state/counter.reducer";
-import { MyCounterComponent } from './my-counter/my-counter.component';
-
-@NgModule({
-    declarations: [
-        AppComponent,
-        MyCounterComponent
-    ],
-    imports: [
-        BrowserModule,
-        StoreModule.forRoot({count: counterReducer})
-    ],
-    providers: [],
-    bootstrap: [
-        AppComponent
-    ]
-})
-export class AppModule {
-}
