+++ web/src/app/components/my-counter/my-counter.component.html
@@ -1,0 +1,7 @@
+<button (click)="increment()">Increment</button>
+
+<div>Current Count: {{ count$ | async }}</div>
+
+<button (click)="decrement()">Decrement</button>
+
+<button (click)="reset()">Reset Counter</button>
