+++ web/src/app/components/my-counter/my-counter.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { MyCounterComponent } from './my-counter.component';
+
+describe('MyCounterComponent', () => {
+    let component: MyCounterComponent;
+    let fixture: ComponentFixture<MyCounterComponent>;
+
+    beforeEach(async () => {
+        await TestBed.configureTestingModule({
+            declarations: [MyCounterComponent]
+        })
+            .compileComponents();
+    });
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(MyCounterComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});
