+++ web/src/app/components/page-not-found/page-not-found.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    template: '<h1>Page Not Found!</h1>'
+})
+export class PageNotFoundComponent {
+}
