+++ web/src/app/state/counter.actions.ts
@@ -1,4 +1,4 @@
-import {createAction} from "@ngrx/store";
+import { createAction } from '@ngrx/store';
 
 export const increment = createAction('[Counter Component] Increment');
 export const decrement = createAction('[Counter Component] Decrement');
