+++ web/src/app/state/counter.reducer.ts
@@ -1,5 +1,5 @@
-import {Action, ActionReducer, createReducer, on} from "@ngrx/store";
-import {decrement, increment, reset} from "./counter.actions";
+import { Action, ActionReducer, createReducer, on } from '@ngrx/store';
+import { decrement, increment, reset } from './counter.actions';
 
 export const initialState = 0;
 
@@ -10,7 +10,7 @@ const _counterReducer: ActionReducer<number> = createReducer(
     on(reset, () => 0),
 );
 
-export function counterReducer(state: number, action: Action): number {
+export function counterReducer(state: number | undefined, action: Action): number {
     return _counterReducer(state, action);
 }
